Moved font spacing functionality to a different file
